# Conflicts:
#   Documentation/Chapters/Sponsoring/Index.rst

+++ ext_tables.sql
@@ -41,7 +41,7 @@ CREATE TABLE tx_gridelements_backend_layout (
 CREATE TABLE tt_content (
 	colPos int(11) DEFAULT '0' NOT NULL,
 	layout varchar(255) DEFAULT '' NOT NULL,
-	backupColPos smallint(6) DEFAULT '-2' NOT NULL,
+	backupColPos int(11) DEFAULT '-2' NOT NULL,
 	tx_gridelements_backend_layout varchar(255) DEFAULT '' NOT NULL,
 	tx_gridelements_children int(11) DEFAULT '0' NOT NULL,
 	tx_gridelements_container int(11) DEFAULT '0' NOT NULL,
